dnsmasq: add listen_address parameter
This patch adds a parser for the uci representation of
dnsmasq's "-a | --listen-address" option.

In summary, this option forces dnsmasq to listen on the given IP address(es).
Both interface and listen-address options may be given, in which case the
set of both interfaces and addresses is used.

Note that if no interface option is given, but listen_address is, dnsmasq
will not automatically listen on the loopback interface. To achieve this, its
IP address, 127.0.0.1 or ::1 must be explicitly given as a listen_address
option.

This option is useful for ujailed dnsmasq instances, that would
otherwise fail to work properly, because listening to the
"This host on this network" address (aka 0.0.0.0 see rfc1700 page 4)
is not allowed.

Signed-off-by: Christian Lamparter <chunkeey@googlemail.com>